1. contracts to java 2. fix transaction data conversion errors
