+++ src/test/java/irita/sdk/evm/Counter_sol_Counter.java
@@ -1,4 +1,4 @@
-package irita.sdk;
+package irita.sdk.evm;
 
 import java.math.BigInteger;
 import java.util.Arrays;
